requirements.txt
logger
azure-servicebus